CLOSED - # 98: LLRP _C1G2ReadOpSpecResult byte->short conversion error https://trac.rifidi.org/rep-external-trac/ticket/98
Also fixed a problem with a type conversion when checking to see if C1G2AirProtocol is the parameter for an AccessSpec

+++ rifidi/trunk/org.rifidi.emulator.reader.llrp/src/org/rifidi/emulator/reader/llrp/commandhandler/LLRPAccessControl.java
@@ -124,9 +124,9 @@ public class LLRPAccessControl {
 			accessReportTrigger = specIn.getAccessReportSpecParam()
 					.getAccessReportTrigger();
 		}
-
+		
 		// If the protocol is C1G2
-		if (specIn.getProtocolId() == 1) {
+		if (specIn.getProtocolId() == (byte)1) {
 			C1G2TagSpec tagspecIn = (C1G2TagSpec) specIn
 					.getAccessCommandParam().getTagSpecParam();
